Added PR button to Tab
