+++ src/app/(layout)/(dashboard)/workspaces/[workspaceId]/projects/[projectId]/client.tsx
@@ -74,18 +74,6 @@ export const ProjectIdClient = () => {
     if (!project) return "";
     return `/workspaces/${project.workspaceId}/projects/${project.$id}/members`;
   }, [project]);
-  const handleCreatePr = async () => {
-    try {
-      await openPr();
-    } catch (error) {
-      console.error("Error creating pull request:", error);
-      toast.error(
-        typeof error === "string"
-          ? error
-          : "You have to push to the specified branch first.",
-      );
-    }
-  };
 
   const handleFileUpload = async () => {
     try {
@@ -182,8 +170,8 @@ export const ProjectIdClient = () => {
           <DropdownMenu>
             <DropdownMenuTrigger className="ml-2" asChild>
               <Button variant="outline" size="default" className="items-center">
-                <EllipsisVertical className="size-4" />
                 <p className="text-sm">Actions</p>
+                <EllipsisVertical className="size-4" />
                 <span className="sr-only">Open menu</span>
               </Button>
             </DropdownMenuTrigger>
@@ -196,14 +184,6 @@ export const ProjectIdClient = () => {
                 >
                   <UploadIcon className="mr-2 size-4" />
                   Upload Readme
-                </Button>
-                <Button
-                  className="w-full justify-start bg-slate-200 text-black hover:bg-slate-300"
-                  onClick={handleCreatePr}
-                  variant="default"
-                >
-                  <GitPullRequestCreateArrowIcon className="mr-2 size-4" />
-                  Create Pull Request
                 </Button>
                 <Button
                   className="w-full justify-start"
